+++ src/attacker.cpp
@@ -1,7 +1,7 @@
 //
 // Created by 温和 on 2018/10/1.
 //
-#include "attacker.h"
+#include "/Users/wenhe/code/Pokemon/include/attacker.h"
 
 AttackerCreatrue::AttackerCreatrue() {
     this->setName("Attacker");
@@ -12,7 +12,7 @@ AttackerCreatrue::AttackerCreatrue() {
 void AttackerCreatrue::upgrade() {
     if (this->getLevel() < LEVEL_TOP) {
         this->setLevel(this->getLevel() + 1);
-        this->setHp(this->getHp() + this->randomGrow(ATTACKER_GROW_HP));
+        this->setHp(static_cast<int>(this->getHp() + this->randomGrow(ATTACKER_GROW_HP)));
         this->setStrength(this->getStrength() + this->randomGrow(ATTACKER_GROW_STRENGTH));
         this->setDefense(this->getDefense() + this->randomGrow(ATTACKER_GROW_DEFNESE));
         this->setSpeed(this->getSpeed() + this->randomGrow(ATTACKER_GROW_SPEED));
@@ -22,7 +22,7 @@ void AttackerCreatrue::upgrade() {
 void AttackerCreatrue::initAbility() {
     static std::default_random_engine e;
     static std::normal_distribution<float> n(0, 0.2);
-    this->setHp(ATTACKER_INIT_HP + round(n(e) * 10) / 10);
+    this->setHp(static_cast<int>(ATTACKER_INIT_HP + round(n(e) * 10) / 10));
     this->setStrength(ATTACKER_INIT_STRENGTH + round(n(e) * 10) / 10);
     this->setDefense(ATTACKER_INIT_DEFNESE + round(n(e) * 10) / 10);
     this->setSpeed(ATTACKER_INIT_SPEED + round(n(e) * 10) / 10);
