+++ src/tanker.cpp
@@ -1,7 +1,7 @@
 //
 // Created by 温和 on 2018/10/7.
 //
-#include "tanker.h"
+#include "/Users/wenhe/code/Pokemon/include/tanker.h"
 
 TankerCreatrue::TankerCreatrue() {
     this->setName("Tanker");
@@ -12,7 +12,7 @@ TankerCreatrue::TankerCreatrue() {
 void TankerCreatrue::upgrade() {
     if (this->getLevel() < LEVEL_TOP) {
         this->setLevel(this->getLevel() + 1);
-        this->setHp(this->getHp() + this->randomGrow(TANKER_GROW_HP));
+        this->setHp(static_cast<int>(this->getHp() + this->randomGrow(TANKER_GROW_HP)));
         this->setStrength(this->getStrength() + this->randomGrow(TANKER_GROW_STRENGTH));
         this->setDefense(this->getDefense() + this->randomGrow(TANKER_GROW_DEFNESE));
         this->setSpeed(this->getSpeed() + this->randomGrow(TANKER_GROW_SPEED));
@@ -22,7 +22,7 @@ void TankerCreatrue::upgrade() {
 void TankerCreatrue::initAbility() {
     static std::default_random_engine e;
     static std::normal_distribution<float> n(0, 0.2);
-    this->setHp(TANKER_INIT_HP + round(n(e)));
+    this->setHp(static_cast<int>(TANKER_INIT_HP + round(n(e))));
     this->setStrength(TANKER_INIT_STRENGTH + round(n(e) * 10) / 10);
     this->setDefense(TANKER_INIT_DEFNESE + round(n(e) * 10) / 10);
     this->setSpeed(TANKER_INIT_SPEED + round(n(e) * 10) / 10);
